Add «Wakeup grace time» timer
The value of "input_datetime.wakeup" is automatically set to the
appropriate value for tomorrow at today's trigger time. This causes an
issue when the wakeup time for tomorrow is later than it is for today
(e.g. on Friday morning 08:00 we set a wakeup time of 09:00 for
Saturday morning): All wakeup routines will trigger again on the same
day at the later time (i.e. Friday at 08:00 *and* 09:00).

Adding a 6 hour grace time "timer" to prevent this issue: If the wakeup
routines have triggered, they will not trigger again for the next 6
hours.