Use U+FE0F to force emoji presentation
For some emojis (e.g. 🌡 and 🌪) some programs on Windows 10 (seeing it
in Firefox and in VS Code) refuse to use a proper emoji and instead
fallback to an "old-fashioned" black-and-white representation.

By using the "Variation Selector-16" (U+FE0F) it is possible to force
the use of an emoji – did this for all the offending emojis and for
some (where I had found an alternative one) revert back to my original
preferred emoji.