Enhance the "Daytime" lighting automations
Made them fully dependent on the new "input_select.twilight" instead of
on the combination of solar elevation and "input_boolean.nighttime".

Only turn the lights in the Living Room and the Kitchen off if they are
still in the "Daytime" scene (to prevent turning off lights that were
intentionally turned on by an occupant).

Finally, when switching states from "dawn" to "daytime", delay the
automation for 60 seconds to prevent us from conflicting with the
regular "lights off at dawn" routine.