Refine and split up group definitions
Instead of single groups.yaml file, use the contents of the groups/
folder. It doesn't appear possible to use multiple "group:" statements
in configuration.yaml, so this'll have to do as an alternative...

The definition "group.everyone" was removed from Git (it's now in the
excluded "groups/private.yaml" file) as it has privacy implications as
it contains the names of the household members.

Apart from that, "group.everyone" was renamed to "group.family". The
existing naming was somewhat confusing, as many of the automations
effectively read "everyone home" whereas most of them meant (and did
trigger on) "anyone home". By renaming the group to "Family" the
semantics make a bit more sense...