Fix edge-case in the Hue scene heuristic
When a *zone* transition causes a room to transition from "off" to "on",
the room reports its last active scene (which isn't actually
representative of its current state). Best way to handle this is to
ignore the transition for the room (the same already happens when a
zone transition causes a room to transition from one "on" to another
"on" state).