Separate out Aarlo automation(s)
Take a snapshot every two hours so the camera previews always show a
more or less up-to-date view.

Add a second automation that stores a snapshot of the back garden daily
at dawn a dusk (could be nice to turn this into a timelaps at the end of
the year). In order to be able to store the snapshot in a sensible
place, allow the "/config/var" location to be written to.