Update/fix the alarm system
Instead of relying on A(a)rlo, there's now a custom alarm control panel
for the alarm system. Arlo (i.e. the cameras) follows this system. If
Arlo goes offline for whatever reason, only control over the cameras is
lost – the remainder of the system remains functional.

Added an independent (Zigbee) alarm siren to replace the one in the
Arlo baseunit.

The Arlo alarm control panel can (for convenience) be armed/disarmed
without an alarm code. To prevent abuse, its controls are hidden from
the Lovelace UI where there is an "active alarm" ongoing.

As opposed to the previous "ad-libbed" implementation, the current one
has been tested to (mostly 🙂) work as intended.