Replace Eurotronic with Tado radiator valve
The Eurotronic (Zigbee) one worked fine, but the Tado has the benefit
of being fully (and easily) integrated with the central thermostat. It
can thus activate the boiler when heat is required. Getting that to work
with the Eurotronic unit would have been a lot of hassle (and would have
probably caused the "intelligence" in the main Tado unit to start
fighting against the automations required).