Add automations for the Kitchen close-in boiler
Switch it off during the night and when no-one is home. Not a major
source of energy savings (especially while working from home) but why
not... 🙂

Includes a provision to ensure the housekeeper has hot water even when
no-one is home and occasional heating when away for longer periods of
time (i.e. vacation) to prevent a Legionella hazard.