Replace Kitchen counter led-strip controller
The new one provide smoother transitions, but it's still "stubborn" from
time to time (not responding to commands). Instead going into an endless
search for a better controller, I've added a little bit of scripting to
ensure transitions in the Kitchen are always fired twice – this seems to
virtually prevent the issue.

Also tweaked the Hue scene heuristic a bit (slightly changed the Kitchen
scenes) and made it more resilient.