Bathroom fan automation (humidity-based; naive)
Switch on bathroom fan when humidity goes above 60%; switch it off when
humidity is below 60% for more than two hours. Naive approach that
(based on the last 10 days of data) should work more or less as well as
manually switching the fan on and off...

Added a derivative sensor for the Bathroom humidity in the hopes that
looking at the derivative makes a more intelligent approach possible
in the future: If humidity goes up quickly, switch the fan on (shower in
use). If the fan is on and the humidity decrease plateaus, switch off
the fan (we've reached the best possible level – regardless of what the
actual humidity level is).

When the outside humidity is very high, it's (probably? – don't know
actually 🙂) going to be difficult to get the bathroom to go below 60%.
As such, the intelligent approach seems better. An alternative could be
to look at outside humidity, but that turns it from a local/self-
contained problem into something totally different...