Make motion lights dependent on light-level
Some rooms/places have sufficient natural light during the day; don't
switch on the lights of that's the case.

Additionally added two template sensors: One to average the light-level
in the Bathroom (from two motion sensors) and another one to average
the light-level between the Bathroom average and the Wardrobe (on the
other side of the house) in order to approximate the light-level in the
upstairs hallway.