Add "Time of Day" template sensor
Aimed at replacing the existing helper/automation combination with
something more concise. Include a secondary sensor which numerically
tracks the primary sensor. This is intended to replace the "Nighttime"
boolean-helper which is used to shade graph backgrounds when it's dark
outside.

Furthermore, reduced the period which is considered dawn/dusk from a
solar elevation of [-6.0, 2.0] to [-4.0, 2.0]. This shifts the regular
civil twilight [-6.0, 0.0] into something more appropriate for our
location (i.e. consider it "dark" outside a bit quicker).