Enhance Attic heating automations
Prevent tado° from (potentially) switching on the heating in the Attic
when it transitions from Away- to Home-mode. Prevent a tado° schedule
transition from unintentionally switching off the Attic heating (which
allows for a more sensible schedule in the Attic without hurting
functionality too much).

Ensure the Attic radiator switches off when the main (Living room)
thermostat switches off (as a fail-safe).