Track average daily gas and energy costs
Divide the current and previous month's costs (generated from the
Utility Meter) by the day of the current month and the total number of
days in the previous month respectively to get to a daily average for
the current and previous months.
These numbers offers a more sensible comparison to the "Today" and
"Yesterday" that were already being tracked.