(Ab?)use Utility Meter to track select appliances
Similar to the DSMR implementation, the Utility Meter (in combination
with a set of smart-plugs) can be used to track consumption (and costs)
of a set of major appliances.

For the cost approximations I'm not bothering too much with the
difference between the peak and off-peak energy rates (as that would
unnecessarily complicate things).