Improve Bathroom fan automation
Instead of acting on absolute values (of relative 😇) humidity, act on
its derivative. This has the advantage of making the trigger independent
of the baseline humidity, which is especially useful in ensuring the fan
switches off when it has no more impact (i.e. it isn't lowering the
humidity any further; on rainy days a baseline relative humidity of
above 40% – the previous turn-off trigger – is not uncommon).