Explicitly specify all state-transitions
In such a way that all of them have a "from"- and "to"-state. This
prevents unintended behaviour when entities transition out of the
"unknown" or "unavailable" states.
Reordered all statements to ensure the "from"-state is always listed
first (makes searching a bit easier).

In two cases the "from"-state is intentionally omitted in the hopes
that it will prevent switches getting stuck in "on"-state upon either
Home Assistant or deCONZ restart.

Finally, it appears to alarm control-panel (i.e. "Home Alarm") always
keeps its proper state. So in order to keep the alarm optimally
"trigger-happy" its transitions are also not explicitly specified.