Further refine House Alarm automations
Add an extra set of automations that enhance alarm behaviour when it is
in "Armed away"-mode. Things that don't make sense (or would be overly
"aggressive") when there's someone in the house.

Provide an overview of the sensor(s) that triggered the alarm in the
alarm notifications.

Improve interaction between the "outside door open at night" automation
and the alarm automations.