Further split up simple-thermostat in Lovelace
There used to be two: One controlling Tadoº via HomeKit and one via its
online service (which is hidden if the online service is unavailable).
Visually the cards are merged into a single card.

The HomeKit card is now itself split up: The header (with all the
"faults" sensors) and the temperature controls. With the thermostat
switched off in Summer, the temperature controls don't make sense (and
they start showing invalid date). So, when the thermostat is off, don't
show them.