Improve Front/Back door open maximum brightness
Turns out there was a failure-mode when the door that triggered the
initial run of the automation got got opened and closed while the
automation was running _and_ the original state of the light was "off".
This would cause the light to remain on indefinetely...

Split up in two separate automations to prevent this issue.