Allow Attic hallway motion-sensing to be paused
Manually operating the Dimmer Switch pauses motion-sensing for half
an hour (or until the light is manually turned off). After half an hour
the lights are turned off and become motion activated again. Prevents
frustration when doing chores in a corner of the Attic... 🙂

Additionally, the "Dim up" and "Dim down" buttons of the Dimmer Switch
now cycle through the two gangs of the smart-relay (one on; two off
– one off; two on – both on).