Broader implementation of Arlo "Arm Home"
When auto-activating the alarm daily at 22:30, use "Home" instead of
"Night" mode. Only the front camera is auto-activated at a fixed time,
the back camera now gets activated when everyone is either asleep or
away.

Furthermore did some gold-plating in Lovelace: The camera feed now has
a coloured border (green/red) indicating whether the camera motion-
detection is active. As this is information is not directly provided by
A(a)rlo, we infer it from the alarm control panel state.