Improve companion app notifications
Use critical notifications for time-sensitive and/or critical (😇)
stuff. Allow for notifications to only be send to devices that are in
the house.

The notification scripts are currently duplicated for convenience. Might
be a way to prevent this duplication (but as of yet I'm unsure whether
that'll make their overall usage easier/clearer).