Various Aarlo tweaks
When following the Home Alarm state, wait for Aarlo to be connected.
This could lead to some unwanted race conditions, but it is generally
better than not doing anything in case Aarlo is disconnected when a
Home Alarm state change occurs

Also, update the Aarlo captures when the Home Alarm is armed or
disarmed.