Integrate "rain hour" instead of "rain rate/hour"
It seemed reasonable to assume that integrating the rain rate (mm/h)
would lead to the total rainfall. It appears though the number greatly
overshoots reality... Based on the data provide, the integral is correct
(so it's not an issue with the computation).

It appears better to use "rain hour" which, instead of providing a rate,
simply provides data points indicating how much rain (in absolute terms)
fell in the previous period (i.e. between n and n-1). This seems to
closely track the "day rain" number provided by WeeWX.

Also revert the prior change to InfluxDB and start pushing "rain hour"
there again too (as it appears it is the most reliable way to calculate
rain totals).