Tweak Garden lights automations
Instead of always switching the lights off at 22:30 (which in the middle
of Summer is actually _before_ the lights get switched on), switch them
off at 23:00 but wait for the lights to have been on for at least 90
minutes (making them stay on longer in Summer).

Also, toggle the Front and back door lights on together with the
interior lights (i.e. a bit earlier) and switch them off independently
of each other (Front door at 22:30 – when the front camera is
switched on – and the Back door together with the Garden).