Refine lighting during alarm situations
When coming home late at night (with the lights off) some of the inside
lights will now come on automatically either directly when disarmed or
after opening a door (depending on what is most appropriate).

The lights also come on after manually disarming a triggered alarm and
the lights switched on during (pending) alarm situations try to switch
themselves off a bit more intelligently as well.