Somewhat unorthodox tweak to Hue scene heuristic
The Kitchen scene tracks the Kitchen counter & cupboards scene. Due to
the way the heuristics is set up, it is possible for the former to
change, without the latter to update its state. This breaks the scene
heuristic for Kitchen.

Found a slightly creative way to force an update; seems acceptable for
now. Let's see what the future holds...