Add automations for a VOCOlinc aroma diffuser
Basic stuff for now – turn it at the start of the evening for 30
minutes if someone is home; turn it off if it's been on for more than 90
minutes and if everyone leaves.

Make the light (sort of) follow the Hue scene in the Living room (but
only when the diffuser is on).