Enhance Attic occupancy detection
Use the "occupancy" of the office chairs as a proxy for the Attic
occupancy, in combination with the state of the lights.

The lights are leading – when the office chairs appear to be empty for
a while a notification is raised asking whether the lights can be
switched off (since it appears no-one is in the Attic). Switching them
off would then also toggle the Attic to "unoccupied".

Depending on how well this work (i.e. how often this notification is
raised), it might (or might not 😉) at some point be possible to rely
solely on the office chairs.