Reduce complexity of Attic airco automations
The original setup was overkill – might revisit the approach later on
(when another airco gets installed in the bedroom).

For now there's a schedule defined in the tadoº-app (airco on weekdays
from 09:00 to 21:00 unless in "Away"-mode). The automations are only
concerned with switching the schedule on and off based on occupancy in
the Attic.