Apply fixes for inconsistent Hue behaviour
Mainly caused by third-party devices (Gledopto and Innr). Switch to a
reactive approach where, instead of going out of our way to prevent the
unwanted behaviour, we detect the unwanted behaviour and correct for it.
The resulting approach is slightly more complicated, but it is limited
to a single place (instead of needing to be applied whenever we're
interacting with the third-party devices).