Correct for Gledopto on/off behaviour
For the "Lights out"-script, instead of unconditionally switching off
all rooms at the end (which causes some issues with the Gledopto
controllers), only switch off the rooms that still report their state as
"on".

For the "lights of at dawn"-automation, only run it if they lights are
actually on (better approach to begin with) _and_ explicitly check the
state of the Kitchen lights (to account for the Gledopto controllers).