Consider number of lights in Hue scene heuristic
Instead of only looking at the state of the Hue room/zone (i.e. group),
also consider the number of lights switched on. This prevents a host of
edge cases as it's impossible to distinguish between a single light
switched on and the entire room switched on by purely looking at the
data reported by the Hue room/zone.