Upgrade to Home Assistant 2017.7.2
As the hourly gas consumption reading was removed from the DSMR
integration, some minor modifications were required (all for the
better). The setup of gas consumption now aligns more closely with that
of power consumption.

An approximation of the old hourly measurement was reintroduced using
the Derivative integration. It's primarily needed for graphing in the
Lovelace dashboard; for convenience it's also still pushed into Influx
(although the cumulative gas consumption measurement should be preferred
there).