Initial Gammu / SMS notification support
Gammu runs in a separate container (using sms2mqtt) and is controlled
via MQTT. This offers a somewhat more flexible solution than controlling
the serial device directly via the Home Assistant and its SMS
notification integration (e.g. if anything goes wrong with the serial
device only the sms2mqtt container needs to be restarted and not Home
Assistant as a whole).