Further refine Gammu / SMS notifications
Use trigger-based template sensors instead of MQTT sensors (which
afford much greater flexibility). Cope with the fact that the sensor
sensor states/values don't always change (as it always represents the
state of the last message received/sent).

Add the capability to send SMS notifications (aligned with the regular
persistent notifications so it's easy to combine both) – including the
ability to send to a custom phone number.

Add an "authenticated command"-sensor which only updates when commands
are received from authorised phone numbers and include the preshared
secret key. Provide responses for two basic SMS commands ("ping?" and
"status").

Finally, send a test SMS notification as part of the monthly alarm test.