Improve alarm pending/sounding notifications
For the pending notification: Instead of listing all sensors that are
"on", only list sensors that are "on" _and_ were changed in the last
five minutes. This much more accurately identifies exactly which sensor
tripped the alarm.

For the sounding notification: List _all_ possible trigger sensors that
changed in the last five minutes. Less accurate than the pending
notification, but better than not listing anything at all...