Better handle back gate open when away
Initial actions are still the same, but if the alarm gets disarmed
within two minutes we assume a false positive and attempt to gracefully
recover from the alert state created.

As false positive occur regularly, only send an SMS messages if the
alarm has been armed for more than 12 hours. This prevents an SMS from
being send during "normal" days still triggering in case we are, for
example, on vacation (at which point receiving such a message might
also be more useful).

Removed the staggered turning on of lights. Unnecessarily complex and
not really noticeable to begin with (by the time you had the door open
and were looking into the garden, all lights were already on).

Finally, blocking the automations for a couple minutes prevents a
potential burglar from emptying out the SMS credit (or the rate limit
for smartphone notifications) by opening and closing the back gate
repeatedly... 😃