More consistent alarm-related light behaviour
If someone enters the Garden when no-one is home, all downstairs lights
are switched on. If subsequently nothing happens, the light are switched
off at dawn (the notification send has a reminder that the downstairs
lights were switched on).

When the alarm gets triggered, switch on the lights in the Garden. If
the alarm subsequently gets disarmed by someone in the house, switch
on the downstairs lights (and keep the Garden lights on) – affords good
visibility in the aftermath of disarming the alarm. If the alarm is
disarmed remotely, switch off the Garden lights – all lights are off
again (since no-one is home there's not much added value in switching on
the lights).