Improve doorbell Arlo snapshot behaviour
When the doorbell rings and no-one opens, instead of outright dismissing
the notification, replace it with a notification containing the latest
Arlo capture (showing who's at the door). Doing this right away is not
feasible as we want to send the doorbell notification as fast as
possible (and thus cannot wait for the Arlo capture to update).

Furthermore, refine the regular/time-based Arlo capture to not race
against the doorbell capture (instead always waiting a couple minutes
before overwriting it). This ensure the notifications contain the
correct capture.

Finally, store the snapshots generated when the doorbell rings. Could
be interesting input for a future machine learning model intended to
identify what is happening at the front door...