Simplify/generalise alarm lighting behaviour
More aggressively switch on lights when the alarm is either pending or
triggered – switching on lights is a good deterrent (and doesn't impact
the image quality of Arlo's night vision).

Reduce the amount of exceptions in place for the "Armed away" situation;
generalise most of it towards all armed states. Turn off more lights as
part of the regular "lights off at dawn" automation as to prevent having
to keep track of this as part of the alarm automations.

Finally, instead of disabling the motion lights when the alarm is
"Armed away", _enable_ them when the alarm is either disarmed or "Armed
night". Under all other circumstances the motions lights shouldn't
operate (either because no-one is home, or because during an active
alarm situation they are controlled by other automations).