Store more Arlo snapshots
Instead a snapshot of the back garden at dusk and dawn, store back and
front garden at dusk, solar noon, dawn and solar midnight. This provides
baseline images with a variety of lighting conditions that might be
useful in training a model to distinguish between "everything's okay"
and "something's up"...

Additionally, download a camera snapshot when a (motion-activated)
recording is made.