Fix persistent notification wait-for-trigger
It appears waiting for "script.turn_on" will trigger if the provided
guard condition (service_data.variables.XXX) is not present in the
actual call to "script.turn_on" (i.e. trigger _only_ on a certain tag
would only work if the tag was provided in the script call, without a
tag provided it would always trigger).

Easier (and more concise) solution appears to be to trigger on
"persistent_notification.create" instead. This always occurs (regardless
how the actual persistent-notification script is called ) and always
has a tag provided.