Enhance "failed devices" sensor/notification
Turns out there are quite some entities not linked to a native device
(either because they're template-entities, or because their integration
itself doesn't provide native devices).

Greatly extended the template sensor to also include all entities not
linked to a device (if a device is present, behaviour is unchanged).
When no device is linked, attempt to manually group the entities into
a "device". Only if that's also not possible, report on the actual
individual entities.

The greatly increases the reach of the sensor, without creating an
over-long list of individually failed entities. Drawback is that there's
quite a bit of logic required to group the entities, but that currently
seems to be outweighed the utility of this sensor/notification.