Use orientation to infer window state
Use the "orientation"-attribute from vibration sensors to infer a more
detailed window state (not just open/close) for several tilting windows.

Update the "rain notification" to only trigger a TTS notification when
the Attic window is fully open. Don't bother with the Bedroom windows
anymore (not a lot of potential for raining in; it appears not feasible
to use the vibration sensor's orientation in this case).