Attic radiator tweaks
Apply some of the improvements made to the Attic Airco to the radiator
as well (using occupancy instead of lights only; switch off the
radiator when the window is opened; various minor tweaks).

Since the Attic radiator now also controls the boiler, there's no need
to keep the Living room thermostat artificially high during weekdays
(as to ensure the boiler is on). As such, the "on" temperature threshold
was lowered to 16ºC.