Upgrade to Home Assistant 2021.10.6
Includes a handful of fixes related to stricter template evaluation.
Only in places where errors were encountered as part of the happy flow;
in exceptional cases the new behaviour (not rendering the template and
writing errors to the log) is probably preferred...

Furthermore, the deCONZ integration now supports alarm devices (more)
natively. Instead of having "switch"-entities, the sirens are now
"siren"-entities (whose names are indicative of their placement in the
house, so these have been obscured from view).