Further refine the wakeup alarm
A cleaner approach relying on a separate script and a (snooze) timer
instead of on the state of the wakeup alarm automation (which could only
be indirectly inferred).

Using the on-state of the wakeup alarm script and the active-state of
the snooze timer, the various automations involved are now a lot less
tightly coupled.

Bedroom lights can't be toggled anymore (i.e. switched off) while the
alarm is snoozed and if no interaction happens after the first time the
alarm sounds, it will sound again after 5 minutes.