Add second Neato D7 to Lovelace
The Neato-related automations still need to be updated...

My old trusty Neato XV-15 (from 2013) launched itself down the stairs
for the _second_ time last week. Its drop sensors broke a couple years
ago (which was discovered the _first_ time it went down the stairs) and
I've been running with only the magnet strips guarding the staircase.
Worked fine until it didn't anymore... 😉

Even tough damage (to the Neato) is limited, I'd prefer this not to
happen again. So, it got replaced by a Botvac D7 which – in addition to
the XV-15's drop and magnet strip sensors – also has virtual no-go
lines.