Remove time-based "Neato active" sensors
The upstairs sensors become superfluous; the D7 upstairs is "Internet-
connected" and thus provides information on its current state directly
to Home Assistant.

The old model in the Attic is now physically blocked from entering the
hallway (mainly to prevent it from accidentally launching itself down
the stairs) and will thus never trip any motion sensors (which negates
the need for the sensor).