Toggle Kitchen lights on "sustained activity"
If the Counter & cupboards detect motion for a while, it seems safe to
assume someone's working in the Kitchen (and it might thus be useful to
switch on all the lights).

The timings are up for experimentation – as it's dependent on the
initial (motion-activated) trigger for the Counter & cupboard lights
there's not a whole lot of additional logic required (I hope/think 😉).