Further improve Kitchen sustained activity lights
This should prevent the "Kitchen counter & cupboards" lights from
dimming (either immediately, or after their automation runs out) if the
Kitchen as a whole is in "Bright (custom)".

Required some tweaks to the automations that cope with bugs/
particularities in the Hue lights. Some (in hindsight unnecessary)
safeguards were removed from these automations