Improve pre-commit linter script
The exit-code of the linter script represented only the result of the
final check (i.e. a prior failing check would not abort the commit).
Use a generalised approach of filtering irrelevant files (instead of
relying on the particularities of every individual tool). Use a diff
filter to filter out deleted files (instead of explicitly testing if
files exist). Ensure spaces in filenames are properly escaped.