Ensuite lighting automations
If Ensuite is set to 100% brightness, keep it at that level regardless
of what happens with the Living room as a whole (unless the Living room
is switched off).

Update the Magic Cube to toggle Ensuite maximum brightness (switching
between maximum brightness and the "correct" setting based on the Hue
scene in the Living room – instead of just setting it to maximum
brightness).

Add an automation that tries to cope with the fact the Ensuite sometimes
doesn't follow the "on"-state of the Living room as a whole (similar
in goal, but different in implementation to the Gledopto automations
for the Kitchen).

Removed "📄 switches.yaml" and moved the automations contained in to
their respective files inside "📁 lights".