Prepone wakeup alarm based on Albert Heijn ETA
Albert Heijn sends notifications with delivery ETA. We mostly receive
deliveries early in the morning, so the notifications aren't super
helpful (as they mostly arrive while we're still asleep). So...

Parse the notifications and if the delivery ETA is before the
scheduled wakeup alarm, prepone the wakeup alarm. This way, the wakeup
alarm can be set later without running the risk of being asleep when
the delivery arrives. Or at least, that's the theory 🤞