Fix washing machine and dryer notifications
I keep overlooking the fact that "wait_for_trigger" triggers can't be
used in combination with the trigger-condition.

Also, vibration sensor turned out to be a bad choice to detect whether
the washing machine door was opened; switched to a contact sensor. Had
previously tested using a vibration sensor to detect whether the washing
machine was running – that didn't work... I guess the position of the
sensor on the machine matters (apparently, the door picks up a lot more
vibration than the side of the machine does).