Upgrade to Home Assistant 2022.5.5
Resolved an issue with the SNMP integration. Furthermore, upgraded the
Aarlo custom component to its latest beta in an attempt to resolve a
connectivity-issue (which turned out to be unrelated). Switched Aarlo
back to the SSE back-end for the time being (effectively making it
behave as the previous stable release).