Add additional info to Security dashboard
Nabu Casa- and Internet-connectivity (as a visual reminder, SMS
notifications are already in place) and an indication of whether
"Vacation-mode" is active (automatically activates when we're away for
a longer period of time).

The occupancy.yaml-file got too complex to be excluded from Git; names
of persons listed have been moved to secrets.yaml.