Fix Albert Heijn ETA edge case
With the recent Home Assistant update that restores the value of
template sensors on restart, an edge case in the Albert Heijn ETA
automation surfaced: The date of the Albert Heijn ETA wasn't explicitly
check (instead, assumed to be for today as normally the notification
always comes in on the same day). With this recent change to Home
Assistant, a stale ETA value coupled with a HA restart would cause the
wakeup alarm to sound immediately.