Tweak alarm pending/triggered notifications
Sort the list of tripped sensors from newest to oldest (i.e. ensure the
last tripped sensor is at the top of the list) and print the total
number of sensors tripped in the message.

Android (and I assume iOS too) truncate notification messages without
any visual cue, so a long list of tripped sensors would simply be cut
without any indication, leading to confusion. By ordering the list and
printing the number of tripped sensors, such confusion is (hopefully)
prevented in the future...