Tweak persistent and SMS notification scripts
Notifications on Android ignore literal line-breaks (and only display
HTML line-breaks), notifications in Home Assistant appear to display
both. For consistent behaviour (and in order to not have to constantly
mess with literal line-breaks in the YAML files) replace all literal
line-breaks before generating the notifications.

Moved processing of the SMS text-message a bit earlier in the script.
Cleaner and allows the actual message sent to be logged in case of
error.

One of the alert notifications had its plain-text (SMS) and HTML message
swapped...