Replace Vacation-mode sensor with input boolean
As its usage increases, an input boolean makes more sense as this allows
the Vacation-mode to also be manually toggled (useful both when going on
vacation and in case of unexpected behaviour while on vacation).
Toggling the input boolean with a set of automations further increases
its flexibility.

Removed the "Guest-mode" input boolean for now as it hasn't been
implemented yet...