Prusa MINI+ tweaks
Add a notification on print job complete. Tweak the dashboard so it is
mostly hidden when the printer is not printing (and all sensors turn to
"unavailable").

The "poor man's variable" in the template sensor didn't survive a
restart of Home Assistant: It works fine while running, but on restart
the first render of the sensor crashes hard (which hints towards the
fact it's smart to simply duplicate the regular expression, so it is).