Replace NVMe with Corsair GTX flash drive
In an attempt to further troubleshoot the USB disconnects, I've replaced
the Orico NVMe SSD enclosure (RTL9210) housing a Kingston A2000 SSD with
the most expensive USB flash drive I've ever purchased: A Corsair
Voyager GTX (twice the price for half the storage compared to the
Orico/Kingston-combination).

Preliminary testing shows it does actually outperform the A2000 by quite
a bit...

Let's see how the stability holds up. One caveat already though: Its
size is extremely awkward for use in Raspberry Pi 4. It only fits in the
top USB 3.0 port (doesn't fit next to the RJ45-connector in the bottom
port) and in doing so blocks access to the adjacent USB 2.0 port 🫤

For now 3 USB ports is all I need, so I'm gonna spare myself the trouble
of finding (and potentially troubleshooting) a good USB hub for use with
the Raspberry Pi.