Initial groundwork for more "Voice"-features
Mainly aiming towards the TTS-side of things for now: Home Assistant
announcing stuff around the house (in addition to / instead of sending
textual notifications).

For now this just means I have a handful of additional speakers up and
running and want to test their stability. What better way of doing that
than having an hourly "Dong!"?