Add some more elaborate TTS announcements
Centralised a handful of common (and complex) pieces of TTS text in
dedicated template sensors that can easily be sourced in multiple
places.

The "Goodnight" message is now also played in the Living Room when
appropriate. There's a "Good morning" message and a "Welcome home"
message with a common "Home Alarm" summary; the "Welcome home" message
has a further, elaborate, personal welcome (still have to see how well
this holds up in practice).

To allow for the "Home Alarm" summary a set of template sensors are
introduced that count up the number of relevant events while the alarm
is armed.

The "Goodnight" message now also includes the Albert Heijn ETA. To
prevent having to constantly replace "Heijn" with "Heyn" (and have it
pronounce my name properly for that matter) there's a hard-coded
exception in "script.tts_announce" that transforms "ij" in "y".