Replace SNMP with "sysmon-mqtt"
Instead of relying on SNMP to push a (very limited) set of metrics to
Home Assistant, use a simple Bash-script to push them over MQTT instead.

Currently designed for (and only tested with) Raspberry Pi 4 devices
running Ubuntu (or in an Alpine container), but in principle it should
work almost anywhere...

The "sysmon-mqtt.sh" script can be run directly, or started through a
Docker-container for additional convenience.