Simple system monitoring via MQTT (#1)
Instead of relying on SNMP to push (a limited set of) metrics to Home 
Assistant, use a simple Bash-script to push them over MQTT.

The "sysmon.sh"-script can be run directly, or started in a Docker-
container for additional convenience. The Docker-images are build using
a manually triggered GitHub worklfow in this repository. Images are
pushed to both Docker Hub and GHCR.

For all supported devices (with the exception of the DD-WRT units) SNMP
has been replaced with this new approach. The DD-WRT units will be
picked up at a later point (probably around the same time as moving the
main router over to OpenWRT).

Added the new Home Assistant Yellow unit to the "monitoring mix" as
well...