Provide container to run external ADB Server
In an attempt to resolve a Shield TV / Android 11 issue, which (I
realised as part of implementing this) appears to have a totally
different solution. Anyway, the container was already up-and-running,
so why not use it for a while...

It offered a nice opportunity to experiment with s6-overlay v3 and does
offer the benefit of being able to easily connect into the Shield TV via
ADB even when Home Assistant is running.