Add some APT-metrics to sysmon-mqtt
Now registers a Home Assistant Update-entity that reports the number
of APT packages that can be upgraded. Also reports if restart is
required as a result of package upgrades.

Should work on any Debian(-derived) distribution. Was initially using
"check-apt" from Ubuntu's "update-notifier" but that didn't turn out to
be very portable, so instead of that approach I'm now parsing APT's
output directly (more portable, slightly less stable).

Running the script inside a Docker container is now less of an option
(as these APT-metrics cannot be sensibly reported from a container). It
still works (and gracefully degrades) though.