Combinations of "not_from/to" state-triggers...
...don't work as intended (or at least, not as how I would expect them)
to work: The combination of "not_from/to" unavailable/unknown seems to
always ignore one of the two unavailable/unknown-states (and thus still
trigger when coming from "unknown" into a regular state). This mainly
occurs when reloading template entities (which is infrequent, but
still annoying).

For now work around it by removing one of the two in places where it
doesn't matter (triggers more often, without side-effects), or by
removing one of the two and adding in additional guard conditions to
prevent unwanted triggers.